# src/main/java/com/example/teamflow/task/dto/request/TaskStatusUpdateRequest.java
package com.example.teamflow.task.dto.request;

import com.example.teamflow.task.entity.TaskStatus;
import jakarta.validation.constraints.NotNull;

public record TaskStatusUpdateRequest(
        @NotNull(message = "Status is required")
        TaskStatus status
) {
}